Use a dictionary instead of a set in MarksCallback
The dictionary uses each original element as key and value:
When an element is marked, it is tested against the key and the value
(the original element) will be given to the result callback, instead of
the marked object.
This avoids to returned an object with a different type than the original one,
like a SleekXMPP JID instead of a string.