Better message handling in HeraldDiscovery (RPC)
The Herald RPC discovery component now waits for its IMessageListener
service to have been registered and notified to Herald Core before
registering its other services.
This avoids the component to send a "contact" message to a peer while it
doesn't listen to the peer's response ("add") yet.

This was the cause of an invalid state, where a peer knew the presence
of a peer, but didn't register the services it provided.